pyproject.toml
[project]
name = "fitness-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.3",
    "fastapi[standard]>=0.115.14",
    "openai-agents>=0.2.3",
    "python-decouple>=3.8",
    "uvicorn>=0.35.0",
]
